requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
pillow==10.2.0
gtts==2.5.1
gunicorn==21.2.0